New color, new class for header, about style added
I have added a new color class to the LESS file, this is light grey which will be used as the background color for some sections.

I have changed the header tag to the newly introduced header class.

Added in color for active, focus & hover states on the navbar

Added styles for the navbar when the new scrolled class is added

Added styles for the new about section

Added styles for the footer section (copyright and designed by)